u-boot-fit-signature.inc: fix issue with u-boot.imx images
Some machines (colibri-imx6ull-emmc and colibri-imx7-emmc) use
u-boot.imx as the U-Boot target, and this is currently not supported by
the default classes responsible for signing the FIT image and adding the
public key to the U-Boot DTB file.

Fix this by implementing the required logic to save the public key to
the U-Boot DTB file.

Related-to: TOR-3060

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>