imx-boot: update signing logic to align with upstream changes
Due to changes on the meta-freescale layer that were incorporated by
Toradex recently in the TorizonCore repo manifest, the build of the
Secure Boot images for i.MX8 machines (both HABv4 and AHAB) were
failing. This commit tries to solve the build issue by aligning the
signing logic (in our .bbappend) with the updated version of the
imx-boot recipe as present in the new version of the layer.

For reference, the meta-toradex layer was initially at commit:

(a) 3e9ef23d98aa842cf84251a27c9b8dde8925ea61

and then the manifest was updated to use instead:

(b) ad424a73b3f815589fef00e366a33eacab3bfcee

which introduced the changes breaking the meta-toradex-security layer;
soon after it was updated to:

(c) 7625e33cd7fd752ee04c7b9b5ba38b73961979ad

on top of which the fixes are being implemented.

Related-to: TOR-3252

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>