u-boot: Add kernel command-line protection
Here we add a new protection to U-Boot whereby the kernel command line
as passed in the U-Boot environment variable "bootargs" is validated
against information in the device-tree (the one passed from U-Boot to
the OS). This information in turn is supposed to come from an overlay
that would be present on every signed OS FIT image.

Related-to: TOR-3246

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>