u-boot: Add HAB support for iMX6Q and iMX6DL
Add HAB support for iMX6Q and iMX6DL based modules. This includes
the Apalis iMX6 and Colibri iMX6 respectively.

After building the U-Boot and SPL binaries the image is signed using the CST
tools and certificate provided by the user. If FIT image signing
support is also enabled then the binaries are re-signed after
being re-built due to needing to rebuild the U-Boot image with the new
DTB.

These changes extend the process from the iMX6ULL. The bulk of the changes
are for handling the additional SPL binary which must also be signed.

Related-to: TOR-3061

Signed-off-by: Jeremias Cordoba <jeremias.cordoba@toradex.com>