u-boot: disable USB gadget support to make room for HAB (mx8m)
With the vulnerability fix introduced on layer meta-toradex-nxp with
commit 5e45485bdd561b3b1b46b3550447e2e4c5e53761, the SPL increased in
size when built with HAB (before the fix the increase due to HAB was
about 10 kB and after the fix the increase became around 28 kB). Because
of this, building secure boot images for the verdin-imx8mm was failing
due to the SPL size exceeding its limit.

With the present commit, we disable the USB gadget support (which is a
not strictly needed feature on Torizon OS) on the i.MX 8M family to make
room for compiling in HAB support. This is possibly a temporary measure
since the same issue will be investigated/tackled at BSP level.

Related-to: TOR-3312

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>