u-boot: enable CONFIG_FIT_SIGNATURE_ENFORCE
CONFIG_FIT_SIGNATURE doesn't really enforce the U-boot setup to be
correct for booting the FIT images, the DTB might not have all the
proper nodes and it just boots up without any warning.

The CONFIG_FIT_SIGNATURE_ENFORCE was introduced in U-Boot in Jun/2023,
and it enforces the FIT signature mechanism to contain the keys in the
DTB and enforce the nodes to be authenticated without relying on the
"required" node in the DTB.

Enabling this option for U-Boot versions that does't support it should
not be a problem, since the config will be just ignored.

Signed-off-by: Sergio Prado <sergio.prado@e-labworks.com>