classes: introduce tdxref-signed class
This class is responsible for creating a verified BSP reference image
with full chain-of-trust support.

It inherits tdx-signed to sign boot artifacts (bootloader and FIT image)
and tdx-signed-dmverity to boot a signed rootfs image with dm-verity.

The tdx-signed-dmverity.class implementation is currently incomplete,
and for now it just builds and boots the rootfs via a ramdisk image
(required for the dm-verity signature verification).

Signed-off-by: Sergio Prado <sergio.prado@e-labworks.com>