tdx-signed-dmverity.bbclass: mount rootfs only from dm-verity image
Change the rootfs to a verity image.

The verity root hash is added to the ramdisk and checked at boot time
when mouting the rootfs.

The rootfs will be mounted and the system will boot only if the hash in
the ramdisk matches the root hash from the verity image.

Signed-off-by: Sergio Prado <sergio.prado@e-labworks.com>