tdx-enc-handler: add recipe
tdx-enc-handler provides the scripts required to handle encryption on
Toradex modules.

It is based on the concept of encryption key backends, that defines where
the encryption key is stored and how it is managed in the device.

For now, it only supports the 'cleartext' backend, that generates a key
on the fly (at boot time) and stores it in the filesystem. This is very
insecure, but we don't care about it, since this is only for testing
purposes.

The plan is to introduce later support for other backends like CAAM, TPM
and OP-TEE.

Signed-off-by: Sergio Prado <sergio.prado@e-labworks.com>