tdx-signed: enable hardening only on Torizon OS images
Since hardening is supported and tested only on Torizon OS, enable it by
default only when building such OS images. At the moment the hardening
features are not expected to work out of the box with BSP images, so
keeping it enabled will cause more harm than good (as it was the case on
issue #20).

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>