u-boot: change Verdin iMX8MP DRAM banks when building with OP-TEE
When booting with OP-TEE we need more DRAM banks, or the boot will fail
with the following message:

CONFIG_NR_DRAM_BANKS is not enough for SDRAM_2

Verdin iMX8MP configuration is hard-coded with 2 banks. Disable the
configuration so the default one is used, which is 4.

Based on https://patchwork.ozlabs.org/project/uboot/patch/20200709084042.8234-12-peng.fan@nxp.com/.

Signed-off-by: Sergio Prado <sergio.prado@e-labworks.com>