tdx-tezi-data-partition.bbclass: introduce class to create data partition
This class will enable the creation of a data partition in the eMMC for
Easy Installer images.

This will make it possible to have a read-write partition for situations
where the rootfs is read-only (e.g. when secure boot/dm-verity is
enabled).

A few additional variables can be used to configure this feature:

- TDX_TEZI_DATA_PARTITION_TYPE: data partition filesystem type (ext2,
  ext3, ext4, fat, ubifs)
- TDX_TEZI_DATA_PARTITION_LABEL: data partition label
- TDX_TEZI_DATA_PARTITION_AUTOMOUNT: configure whether the data
  partition should be automatically mounted at boot time or not
- TDX_TEZI_DATA_PARTITION_MOUNTPOINT: data partition mount point
- TDX_TEZI_DATA_PARTITION_MOUNT_FLAGS: data partition mount flags (to be
  added to fstab)

Signed-off-by: Sergio Prado <sergio.prado@e-labworks.com>