tdx-optee.bbclass: enable usage when dm-verity is enabled
OP-TEE needs a read-write filesystem (by default, it writes to
/data/tee/). When rootfs signature checking is enabled via the
tdxref-signed class, the rootfs image will be generated using the
dm-verity kernel feature, which is read-only.

In this case, to provide a read-write filesystem to OP-TEE, the
tdx-tezi-data-partition class will be automatically inherited. This
class will create an additional partition in the eMMC and mount it by
default at /data.

Signed-off-by: Sergio Prado <sergio.prado@e-labworks.com>