u-boot: redo the integration of HAB signing for i.MX6/6ULL/7
With commit 5e12dc911d0c541f43aa6d0c046fb87e8b7c1f7e on layer
openembedded-core, the class uboot-sign underwent big changes where,
amongst other things:

- All the doubly nested loops interating through UBOOT_MACHINE and
UBOOT_CONFIG were made to call helper functions to perform the actual
work. This allows one to avoid duplicating those loops when adding code
to tasks by simply prepending or appending directly to the helper
functions.

- The task responsible for assembling a FIT image is always present even
when FIT signing is not enabled and it always runs after U-Boot is
compiled.

Here we leverage both of the above facts to heavily simplify the HAB
signing process. Now we no longer have the doubly nested loops and the
signing is always performed in the same task (uboot_assemble_fitimage)
independent of whether or not FIT signing is enabled; before, the HAB
signing happened on task compile or deploy depending on whether FIT
signing was enabled on not.

Besides the above, we also add a couple of improvements to the CSF
creation script and to the actual signing function (sign_habv4) to help
with maintenance.

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>