u-boot: update logic to detect bad usage of hardening
The original logic intended to detect the build of BSP images and fail
on them if hardening was enabled; this was done by checking the presence
of override "torizon" - this worked only for images built with the
distro of the same name (i.e. on the i.MX8 machines). Now we check both
"torizon" and "torizon-upstream" so that all builds of torizon images
should pass the check while BSP images should fail when hardening is
enabled. Notice that the hardening is disabled by default on those
images so a user would face build problems only if they forcefully
enabled the hardening.

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>