imx-boot-hab: warn users about possibly non-working HABv4 signing
With U-Boot 2024.07 (used on Torizon OS 7), the bootloader container
for the verdin-imx8mm and verdin-imx8mp (both based on HABv4) is
created during the build of U-Boot (via binman), by default. Switching
to the old method where the container is produced by imx-boot is still
possible but (as shown by tests) does not work with U-Boot 2024.07 due
to a combination of increased SPL size and the inefficient container
that imx-boot creates.

With this commit, we show a message letting users know that the
container generation by imx-boot is deprecated and asking them to
switch to the generation by U-Boot.

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>