u-boot-hab: make CSF-related source files specific to i.MX6/7
Add the various CSF generation files (script and templates) to SRC_URI
conditioned on the SoC override. For now we handle only the i.MX6/7
signing so the overrides are for those machines but we are about to add
the handling for the i.MX8M. This work is done in preparation for that
addition.

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>