u-boot-harden: update downstream hardening patches
BSP has switched downstream U-Boot (currently used by the apalis-imx8
and colibri-imx8x) from version v2022.04 to v2024.04 requiring the
hardening patches to be updated, which is done in this commit. Due to
the update one of the patches is no longer required and is thus being
dropped; patches were renumbered thanks to this as well.

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>