u-boot-fit-signature: support FIT image signing alongside binman
This is basically adding support for the kernel FIT image signing when
the bootloader container is produced via binman.

Normally the U-Boot integration by openembedded would generate the FIT
image of the kernel and sign it which at the same time would update the
U-Boot DTB file (u-boot.dtb, produced during the compilation of U-Boot)
with the signing key. Then, in case of i.MX8M SoCs, some extra step
would take the updated DTB and make it part of a bootloader container
which would contain multiple other artifacts such as the DDR training
firmware files.

When binman is employed, the bootloader container is fully produced by
binman and the openembedded integration is not aware of this. The
device-tree blobs placed into the bootloader container by binman are
specified by CONFIG_OF_LIST and the U-Boot DTB is not normally used; due
to this the updated DTB with the signature node for checking the kernel
FIT is not present in the final container.

To fix the above issue, here we take the updated DTB and copy the
signature node to the DTB(s) defined by CONFIG_OF_LIST and then we
re-execute binman to generate and updated bootloader container.

The solution here is ad-hoc but should be robust enough; however,
in the future we should consider more streamlined solutions for FIT
signing to be provided by openembedded or by binman/U-Boot itself.

Related-to: TOR-3539

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>