tdx-enc-handler: add support for saving the encryption key in the disk partition
On systems with secure boot enabled and an encrypted data partition,
there is no place to store the encryption key blob, since the rootfs
will be read-only (because of dm-verity) and one cannot use the data
partition to store the key required to decrypt itself.

To solve this issue, let's make it possible to save the encryption key
blob inside the disk partition to be encrypted, but in a reserved block
outside the dm-crypt partition.

This feature can be enabled by setting the TDX_ENC_KEY_LOCATION variable
to 'partition'.

When enabled, at least one block from the partition is reserved to store
the encryption key blob. In case one wants more blocks to be reserved
for other purposes, the number of blocks to be reserved can be
configured via the TDX_ENC_STORAGE_RESERVE variable.

The key is stored in the partition as a list of null-terminated strings
in the format key=value:

# dd if=/dev/mmcblk2p3 of=/tmp/keyblob.txt skip=30515190 bs=512 count=1
# cat /tmp/keyblob.txt
keyname=tdxenc
keydata=e2204090ffee9159ebaff4e322590c4a4f44f203f84c3411b73a6788c6ae1...
keycsum=29d6ef285d4ea17b0ccf1d94c9fe3957058b2b7733459d33c3ed12f9d031408f

The default behavior is still to save the key as a file in the
filesystem (TDX_ENC_KEY_LOCATION set to 'filesystem'), so this feature will
not break systems that are currently using encryption.

Signed-off-by: Sergio Prado <sergio.prado@e-labworks.com>