imx-mkimage: drop support for HABv4 signing on i.MX8M
Now that we support HABv4 signing on top of the bootlader container
produced via binman (employed on SoCs based on the NXP i.MX8M), drop
support for the generation of the container by the imx-boot recipe
(imx-mkimage tool). If one tries to use the latter method, a
fatal message will show up to let the user know this support was
dropped.

Related-to: TOR-3539

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>