u-boot: display warnings on disallowed commands
With this implementation that is part of the hardening, U-Boot will show
warning messages informing the users when a command that is accepted in
open state would not be allowed when the device is closed. This is
useful to let users know that closing the device would likely leave it
in an unbootable state if any of the commands is part of their boot
script.

Resolves #23

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>