u-boot-distro-boot-harden: make amendments to the boot script
Conditionally apply amendments to the boot script to make it suitable to
use with the hardening feature. Those amendments are applicable to
Toradex BSP images and should be disabled by users having custom boot
scripts, for example. The main modification to the boot script is the
removal of a block involving the "setexpr" U-Boot command which we do
not want to be part of the whitelist.

The amendments are applied only when TDX_AMEND_BOOT_SCRIPT is set to "1"
and the default value of this variable is the same value of
TDX_UBOOT_HARDENING_ENABLE such that when the hardening is disabled, no
amendments are applied by default.

Related-to: TOR-3558

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>