u-boot: validate the root=PARTUUID= bootarg
Validate the root=PARTUUID= bootarg with the goal of adding hardening
support for BSP images; such an argument would be present in the variable
part of the bootargs string in case of BSP images (since the argument is
dynamically set by the boot script). This bootarg is accepted only if
the root= bootarg is not present in the expected bootargs string (inside
the control DTB); this is important because it prevents the root
argument to be modified in cases where want it fixed to some value,
which is the case with Torizon OS that currently selects the root
partition by (a fixed) label.

Notice that the changes introduced here impact basically the hardening
feature which is kept as a series of patches; also, these particular
changes are applicable to both upstream and downstream U-Boot.

Related-to: TOR-3558

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>