imx-boot-hab.inc: stop using variable BOOT_NAME
Upstream project meta-freescale has dropped variable BOOT_NAME since
c30f12b8 (imx-boot: Drop un-necessary variable BOOT_NAME) which was
actually used by this layer to determine the name of the bootloader
binary to be signed. Here we follow along by hard-coding the value that
was previously in that variable.

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>