u-boot: add patches to show message when CLI access is allowed
Minor improvement to hardening (CLI access protection): before, a
message was shown on the U-Boot console only when the device was closed
and the CLI access was disabled by the protection feature. Now we show a
message also when the device is open which is helpful when someone wants
to ensure the protection is compiled in without having to close the
device.

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>