tdx-enc-handler: unmount encrypted partition from all mountpoints
To handle the case where the encrypted partition is mounted at multiple
mountpoints, do the unmount for all associated mountpoints. With Torizon
OS this is important because the encrypted partition might be mounted
more than once in some occasions; for example if the tdx-enc-handler
service is restarted which causes the device-mapper device to be
removed/re-created/mounted by that same service but then mounted a
second time by the usermount service (which currently doesn't have an
easy way to prevent the mount of selected devices). Overall this makes
the solution more flexible because it works even if users add other
mounts for the encrypted partition as a whole or for directories inside
it (bind mounts).

Related-to: TOR-3616

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>