u-boot-harden: simplify command whitelist configuration
Add new OE variables for configuring the whitelist feature:

- TDX_SECBOOT_WL_ALLOW_OPEN_CATEG
- TDX_SECBOOT_WL_ALLOW_CLOSED_CATEG
- TDX_SECBOOT_WL_DENY_OPEN_CATEG
- TDX_SECBOOT_WL_DENY_CLOSED_CATEG
- TDX_SECBOOT_WL_NEEDED_CATEG

With this commit these variables will define the values that go inside
the U-Boot control DTB which in turn controls what commands are allowed
or denied by the feature. Before, a user would need to patch or replace
the control DTB fragment which was a bit cumbersome.

The documentation is also updated to describe the variables and refer to
a more detailed explanation at the Toradex website.

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>