u-boot-harden: add explicit self-overwrite protection
Add new hardening feature to prevent the "load" command from
self-overwriting U-Boot or to write to non-memory regions of the
address space (such as areas held by memory mapped devices).

Before this, the self-overwriting protection relied on the built-in
protection that the "load" command had when the LMB config was enabled
and we simply confirmed that CONFIG_LMB was set. That protection has
shown to be buggy and not fully effective.

Now we have an explicit protection (which still requires CONFIG_LMB to
be set) that additionally limits where the load can happen to the range
of offsets from 0 to 1G-64M from the beginning of the SDRAM (or 512M-64M
on devices with less than 1GB of memory). More details can be found in
the header of the patch file being introduced, namely:

0005-toradex-integrate-load-protection-*.patch

Other patch files have been renamed:

{0005 => 0006}-toradex-cmd-fuse-add-switch-for-quiet-operation-down.patch
{0005 => 0006}-toradex-cmd-fuse-add-switch-for-quiet-operation-upst.patch
{0006 => 0007}-toradex-cmd-fuse-add-readv-command-downstream.patch
{0006 => 0007}-toradex-cmd-fuse-add-readv-command-upstream.patch
{0007 => 0008}-toradex-cmd-setexpr-integrate-protections-downstream.patch
{0007 => 0008}-toradex-cmd-setexpr-integrate-protections-upstream.patch

Related-to: TOR-3771

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>