u-boot: fix built-in protection against self-overwriting
Fix U-Boot built-in protection against overwriting reserved regions via
the "load" command. Tests have shown the built-in protection to be buggy and
ineffective. The patches introduced here try to solve the situation on
both upstream U-Boot and on the downstream version from TI. For NXP
downstream, a similar protection already exists which was introduced
long ago by NXP with the following commits:

- ae2d7bb35f8: MLK-21885-2 lmb: refactor lmb to handle overlap in lmb
  reserve
- 3afc533824a: MLK-21885-1 Revert "lmb: Treat a region which is a subset
  as equal"
- d53a243859c: Revert "lmb: remove overlapping region with next range"

It is worth mentioning that these changes are not strictly required to
ensure the Toradex secure boot solution is effective because the
hardening now has its own more strict protection against self-
overwriting. Thus the patch in the present commit works mainly as an
additional safeguard.

Related-to: TOR-3771

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>