tdx-signed-k3: allow selection between HS-FS or HS-SE devices
Originally, when building a signed image for the K3 architecture, the
target "device type" was always HS-SE (High-Security Security Enforced)
which required customers to run the OTP writer to fuse the device right
after flashing the image with Toradex Easy Installer. This was fine for
users aiming production but for all other cases (like evaluation and
development), it represented a complication that impaired the user
experience.

The present commit tries to improve the situation by introducing a new
boolean variable (named TDX_K3_SECBOOT_TARGET_HSSE_DEVICE) to indicate
if the generated image is supposed to be installed on a HS-FS (High-
Security Field Securable) or on a HS-SE device. The default value of the
variable is 0 which means the target device type will be HS-FS meaning
that an image produced with the default configuration will run normally
on an unfused device.

To avoid confusion we also rename some variables that mentioned the term
HS-SE, as follows:

- TDX_K3_HSSE_ENABLE -> TDX_K3_SECBOOT_ENABLE
- TDX_K3_HSSE_KEY_DIR -> TDX_K3_SECBOOT_KEY_DIR

Since this renaming is a breaking change, we also throw a fatal error if
any of them is set to force users to switch to the new variables.

Besides the variables, also rename some files replacing the term
"hsse" (or "hs-se") by "k3-secboot".

Related-to: TOR-3805

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>