imx-hab.bbclass: throw error if TDX_IMX_HAB_CST_KEY_INDEX is used
Throw an error if the old variable TDX_IMX_HAB_CST_KEY_INDEX is used to
force users to update to the new one, TDX_IMX_HAB_CST_SRK_INDEX, which
is now 1-based. Since the original variable was not documented, we do
not expected many users to be setting it.

Related-to: TOR-3832

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>