u-boot-hab.inc: use libfaketime when generating CSF binaries
When generating the CSF binaries, the NXP Code Signing Tool adds the
current UTC timestamp in string format to them, resulting in files that
are not reproducible. It is possible to circumvent this limitation by
using a library like libfaketime during the build process to trick CST
to always add the same timestamp to the binaries.

Related-to: TOR-3842

Signed-off-by: Lucas Akira Morishita <lucas.morishita@toradex.com>