u-boot-harden: stop build when insecure bootmeths are enabled
When the hardening is enabled, check for the selected "Standard Boot"
bootmeths and stop the build if any bootmeth other than the one called
"script" is enabled.

The "script" bootmeth is used by both Torizon OS and the Toradex
Reference images and is expected to be secure since it basically loads
and executes a script so that each and every command in the script would
need to be allowed by the hardening before they can be executed.

Related-to: TOR-3865

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>