u-boot: properly detect open/closed i.MX95 device
Add support for detecting the current life-cycle state of an i.MX95
device. Before the present commit, the same method used with the i.MX8
was employed with the i.MX95, which didn't work (the result being that
the device was considered as closed regardless of its actual state).

With the appropriate detection being integrated, the various hardening
features that rely on the device state are now supposed to work on the
i.MX95-based devices.

Related-to: TOR-3932

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>